- add xlc test to detect 64 bits architecture
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@44 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/Programs/QHE/QHEBosonsCoulomb.cc
@@ -75,7 +75,7 @@ int main(int argc, char** argv)
   int NbrEigenvalue = NbrEigenvaluesOption.GetInteger();
   int NbrBosons = NbrBosonOption.GetInteger();
   int LzMax = LzMaxOption.GetInteger();
-  int Memory = MemoryOption.GetInteger() << 20;
+  long Memory = MemoryOption.GetInteger() << 20;
   bool DeltaFlag = DeltaOption.GetBoolean();
   double CoulombRatio = CoulombRatioOption.GetDouble();
